requirements.txt
tweepy
python-dotenv
requests
textblob